Newsapi/requirements.txt
newsapi-python
PythonTurtle